requirements.txt
click
networkx
attrs
toposort
git+https://github.com/dhimmel/obonet.git#egg=obonet
